requirements.txt
Flask
cryptography==36.0.1
pycryptodome==3.11.0